refactor: 쿠폰 도메인 내 세부 도메인은 서브 패키지로 정리
