+++ backend/src/main/java/com/woowacourse/kkogkkog/coupon/domain/CouponHistory.java
@@ -1,5 +1,7 @@
 package com.woowacourse.kkogkkog.coupon.domain;
 
+import com.woowacourse.kkogkkog.coupon.domain.event.CouponEvent;
+import com.woowacourse.kkogkkog.coupon.domain.event.CouponEventType;
 import com.woowacourse.kkogkkog.member.domain.Member;
 import java.time.LocalDateTime;
 import javax.persistence.Column;
