+++ backend/src/main/java/com/woowacourse/kkogkkog/coupon/domain/state/CouponStatus.java
@@ -1,6 +1,7 @@
-package com.woowacourse.kkogkkog.coupon.domain;
+package com.woowacourse.kkogkkog.coupon.domain.state;
 
 import com.woowacourse.kkogkkog.common.exception.InvalidRequestException;
+import com.woowacourse.kkogkkog.coupon.domain.event.CouponEventType;
 
 public enum CouponStatus {
