+++ backend/src/main/java/com/woowacourse/kkogkkog/infrastructure/event/PushAlarmEvent.java
@@ -1,6 +1,6 @@
 package com.woowacourse.kkogkkog.infrastructure.event;
 
-import com.woowacourse.kkogkkog.coupon.domain.CouponEventType;
+import com.woowacourse.kkogkkog.coupon.domain.event.CouponEventType;
 import com.woowacourse.kkogkkog.coupon.domain.CouponHistory;
 import com.woowacourse.kkogkkog.member.domain.Member;
 import com.woowacourse.kkogkkog.member.domain.Workspace;
