+++ backend/src/main/java/com/woowacourse/kkogkkog/infrastructure/event/WoowacoursePushAlarmEvent.java
@@ -1,6 +1,6 @@
 package com.woowacourse.kkogkkog.infrastructure.event;
 
-import com.woowacourse.kkogkkog.coupon.domain.CouponEventType;
+import com.woowacourse.kkogkkog.coupon.domain.event.CouponEventType;
 import com.woowacourse.kkogkkog.coupon.domain.CouponHistory;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
