+++ backend/src/test/java/com/woowacourse/kkogkkog/coupon/domain/state/CouponStatusTest.java
@@ -1,17 +1,18 @@
-package com.woowacourse.kkogkkog.coupon.domain;
+package com.woowacourse.kkogkkog.coupon.domain.state;
 
-import static com.woowacourse.kkogkkog.coupon.domain.CouponEventType.ACCEPT;
-import static com.woowacourse.kkogkkog.coupon.domain.CouponEventType.CANCEL;
-import static com.woowacourse.kkogkkog.coupon.domain.CouponEventType.DECLINE;
-import static com.woowacourse.kkogkkog.coupon.domain.CouponEventType.FINISH;
-import static com.woowacourse.kkogkkog.coupon.domain.CouponStatus.ACCEPTED;
-import static com.woowacourse.kkogkkog.coupon.domain.CouponStatus.FINISHED;
-import static com.woowacourse.kkogkkog.coupon.domain.CouponStatus.READY;
-import static com.woowacourse.kkogkkog.coupon.domain.CouponStatus.REQUESTED;
+import static com.woowacourse.kkogkkog.coupon.domain.event.CouponEventType.ACCEPT;
+import static com.woowacourse.kkogkkog.coupon.domain.event.CouponEventType.CANCEL;
+import static com.woowacourse.kkogkkog.coupon.domain.event.CouponEventType.DECLINE;
+import static com.woowacourse.kkogkkog.coupon.domain.event.CouponEventType.FINISH;
+import static com.woowacourse.kkogkkog.coupon.domain.state.CouponStatus.ACCEPTED;
+import static com.woowacourse.kkogkkog.coupon.domain.state.CouponStatus.FINISHED;
+import static com.woowacourse.kkogkkog.coupon.domain.state.CouponStatus.READY;
+import static com.woowacourse.kkogkkog.coupon.domain.state.CouponStatus.REQUESTED;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import com.woowacourse.kkogkkog.common.exception.InvalidRequestException;
+import com.woowacourse.kkogkkog.coupon.domain.event.CouponEventType;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
